int to bigint

+++ Chas-movies/database/migrations/2023_02_14_105505_create_shows_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('genre');
             $table->string('director');
             $table->string('maincast');
-            $table->integer('watchlist_id');
+            $table->unsignedBigInteger('watchlist_id');
             $table->timestamps();
         });
     }
